requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-dotenv==1.0.0
pdfplumber==0.10.3
python-docx==1.1.0
huggingface_hub==0.20.3
python-telegram-bot==20.7
reportlab==4.0.8
Jinja2==3.1.2
Pillow==10.1.0
scikit-learn==1.3.2
requests==2.31.0
aiohttp==3.9.1
